[lib] Separate `serverThreadFrozenDueToBlock` and `clientThreadFrozenDueToBlock`
Summary:
Updating all of the `threadFrozenDueToBlock` callsites on `keyserver` is going to be involved. I started working on it, but it required changing a bunch of functionality in `keyserver`/`lib`/etc.

I will still consolidate the logic in this stack, specifically to avoid the sketchy `CHANGE_ROLE` check for `server` variant of function... but will sequence that after figuring out everything on client.

---

Depends on D11308

Test Plan:
Straightforward refactor, find/replace + `flow`

Reviewers:
tomek, ginsu, michal, kamil