[web] Move worker database globals to worker-database.js
Summary:
Move the `sqliteQueryExecutor` and `dbModule` globals to another shared worker js module and expose functions to get/set them. This simplifies the identity client/ olm api as they can just use the global values. Additionaly I find it clearer in the shared worker code when the "set" operation is more explicit and not just an assignment.

Depends on D11236

Test Plan: Check that the database still works, data is persisted.

Reviewers: kamil, tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D11324