[lib] Refactor fullStateSyncActionType in reducers
Summary:
issue: [[https://linear.app/comm/issue/ENG-6484/update-fullstatesyncactiontype | ENG-6484]]
fullStateSyncActionType always concerns only one keyserver. It is a "specific" action. It has a keyserverID in the payload.

reduceIntegrityStore -> we should not clear the whole store, only remove the threads related to this keyserver

reduceThreadInfos -> we shouldn't remove all threads, only the threads related to this keyserver

reduceCurrentUserInfo -> it should not change currentUserInfo for clients not relying on authoritative keyserver

reduceUserInfos -> it shouldn't be removing user infos for clients not relying on authoritative keyserver

Test Plan: tested that only threads related to the keyserver the action concerns are removed from thread store, and only their hashes are removed from integrity store. Checked that for clients relying on authoritative keyserver currentUserInfo and userStore are being changed

Reviewers: michal, ginsu, kamil

Reviewed By: michal

Subscribers: will, ashoat, tomek

Differential Revision: https://phab.comm.dev/D11313