[lib] Update `clientThreadFrozenDueToBlock` and `getMinimallyEncodedCurrentUser` signatures to include `communityThreadInfo`
Summary:
Updating `clientThreadFrozenDueToBlock` and `getMinimallyEncodedCurrentUser` signatures to include `communityThreadInfo`.

Next step is to update `useExistingThreadInfoFinder` (which is easy) and `threadInfoFromRawThreadInfo` (which is hard and requires recursively updating all the callsites).

Will begin with `useExistingThreadInfoFinder`, and then update `threadInfoFromRawThreadInfo` and callsites, etc and proceed.

---

Depends on D11311

Test Plan: `flow` + careful reading

Reviewers: tomek, ginsu, kamil, michal

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D11317