[keyserver] Remove deprectaed uses of updateTypes.UPDATE_CURRENT_USER
Summary:
issue: [[ https://linear.app/comm/issue/ENG-7205/review-updatetypesupdate-current-user | ENG-7205]]
The two removed files are scripts. They were not used anywhere in the code (they were supposed to be run manually). Now that identity service is the administrator of user data, it doesn't make sense to run scripts that change user data on keyserver.

In `accountUpdater` updates were created back when password was stored in `CurrentUserInfo`. See D1470. Now `accountUpdater` is only used for updating passwords, so there are no changes to `curentUserInfo` it can make

Test Plan: Tested that it is still possible to change the password and then login on this user account from both web and native

Reviewers: michal, kamil, ginsu

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D11327