[web/native] Add content outbound session to olmAPI
Summary:
Part of [ENG-6657 : Migrate session creation code to shared worker](https://linear.app/comm/issue/ENG-6657/migrate-session-creation-code-to-shared-worker)

This corresponds to `contentSessionCreator` from `OlmSessionCreatorContext`. Now that the crypto store is in the shared worker we have to use non-react code to create outbound sessions. This diff adds a new method to `olmAPI` that handles that. On native it just calls out to the native code implementation. On web shared worker the implementation is taken from `OlmSessionCreatorContext` (which is going to be removed in a future diff), only modified to use the global shared worker crypto store instead of unpickling data from redux.

Depends on D11338

Test Plan: Tested that peer to peer communication works with web as the outbound. Didn't test native as it just uses the native code implementation so it works exactly the same as the existing native context provider.

Reviewers: kamil, tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D11339