[web/native] Add getUserPublicKey to olmAPI
Summary:
[ENG-6654 : Migrate getUserPublicKey to shared worker](https://linear.app/comm/issue/ENG-6654)

Continuing the unification effort between web and native APIs this diff adds `getUserPublicKey` to the `OlmAPI`. It will be used in the later diffs.

Depends on D11339

Test Plan: Run this function on native. Pickled the crypto account from native and unpickled it on web. Run this function on web. Checked that the returned value on web is identical to the one from native.

Reviewers: kamil, tomek

Reviewed By: kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D11340