[lib] Split implementations of `[client/server]ThreadInfoFromRawThreadInfo`
Summary:
We compute `currentUser` with `CHANGE_ROLE` approximation on server (for now), and will computer `currentUser` using `communityThreadInfo` on client.

> Where are usages of `clientThreadInfoFromRawThreadInfo`??

Each of the callsites that will need to have `serverThreadInfoFromRawThreadInfo` -> `clientThreadInfoFromRawThread` require a good amount of work. I think it makes sense to split implementations of `*ThreadInfoFromRawThreadInfo` in one diff and handle the callsites separately. Especially because for some callsites we need to recursively update their own callsites.

The callsites that will be updated are:
- `threadInfoSelector`
- `createPendingThread`

---

Depends on D11318

Test Plan:
Updated callsites to use both `server*` and `client*` variants of the function naively and set breakpoint to ensure things worked as expected.

Reviewers:
ginsu, tomek, michal, kamil