[native] Remove connection from persist black list
Summary:
issue: [[https://linear.app/comm/issue/ENG-4933/remove-connection-from-persist-blacklist | ENG-4933]]
`connection` was moved to `kyeserverStore` some time ago, and we make sure it's not persisted using transforms. BEcause it's not a top level value anymore, this code wasn't doing anything.

Test Plan:
this was dead code, because `connection` field is not in top level of state anymore
closed and opend the app a couple of times, no errors showed up, checked that `connection` is not top level but in the keyserver store

Reviewers: michal, kamil, ginsu

Reviewed By: michal

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D11336