[web/native] Move notificationsSessionCreator to olmAPI
Summary:
Part of [ENG-6657 : Migrate session creation code to shared worker](https://linear.app/comm/issue/ENG-6657/migrate-session-creation-code-to-shared-worker)

Simiarly to D11339 this corresponds to `notificationSessionCreator` from `OlmSessionCreatorContext` but without the react-based code (as it doesn't work in the shared worker). The code from `OlmSessionCreatorContext` will be removed in a future diff.

Depends on D11341

Test Plan:
Tested in a later diff but in particular -> checked that the web notifications work and were able to be decrypted using this API to create a session.

Reviewers: kamil, tomek

Subscribers: