[web] Use crypto store and olm from shared worker
Summary:
[ENG-6656 : Update the web code to use new cryptoStore](https://linear.app/comm/issue/ENG-6656)

Replaces the rest of usages of the crypto store in redux in favour of shared-worker-based crypto store.

Note: this diff won't be landed with the future diffs that migrate crypto store data in redux to shared worekr

Depends on D11347

Test Plan:
Tested on web, as on native any changes still call the same underlying methods.
- Logged in to a keyserver in a single-keyserver mode, made sure that the new `cookie` row contained `signed_identity_keys`
- Removed `signed_identity_keys` manually from keyserver db. Reloaded the webapp (which made a new websocket connection). Made sure that after reconnecting `signed_identity_keys` was again filled in, with the same value
- Made sure that the identity client used a non-null device ID
- Made sure that there weren't any error on QR login screen and that the QR login showed up
- Made sure that connection to tunnelbroker worked and that I was able to send messages

Reviewers: kamil, tomek

Subscribers: