[web] Migrate crypto store to shared worker
Summary:
[ENG-6655 : Migrate cryptoStore from redux](https://linear.app/comm/issue/ENG-6655)

Previous diffs modified the code so that they use crypto data in shared worker instead of the data in redux. Now we can migrate the crypto accounts from redux to the shared worker.

Depends on D11348

Test Plan:
Run migration, make sure that the the data that was in redux matches the data in shared worker. Flow.

Reviewers: kamil, tomek

Subscribers: