[web] Stop setting CSAT on initial redux state action
Summary:
issue: [[https://linear.app/comm/issue/ENG-6919/csat-is-set-to-null-after-closing-and-opening-the-web-app | ENG-6919 ]]
From D9122 it looks like `commServicesAccessToken: null` is sent from the keyserver as a result of decoupling web app and keyserver data. And back when keyserver would send all state fields in initial redux state with web app, we were just using the default null.  So it looks like it wasn't supposed to do anything, it's just the default.

But because we were sending `commServicesAccessToken: null`, and initial redux state is being spread in the reducer, the CSAT would get set to `null` when the app was reloaded.
We don't want to stop sending the value from the keyserver though, to not reintorduce [[https://linear.app/comm/issue/ENG-6849/expected-keys-not-rehydrated-[commservicesaccesstoken | ENG-6849]]

Test Plan: Tested that in multikeyserver setup, CSAT is no longer set ot null after app is reloaded

Reviewers: tomek, michal

Reviewed By: michal

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D11350