[web] Deprecate non-worker opaque wasm
Summary:
[ENG-7242 : Deprecate opaque WASM for non-worker web context](https://linear.app/comm/issue/ENG-7242/deprecate-opaque-wasm-for-non-worker-web-context)

D11274, as a part of moving the identity client to the shared worker, introduces a new way to host the opaque WASM file for the webworker that is more inline with the rest of the WASM files. This diff removes the previous WASM file hosted for the browser tab JS context.

Depends on D11351

Test Plan: Made sure that in dev and prod there were no errors in the console after loading the webapp.

Reviewers: kamil, tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D11352