[backup] Revoke holders of replaced log items
Summary:
[ENG-7005 : Handle blob holders of replaced logs](https://linear.app/comm/issue/ENG-7005)
It's possible that backup logs will be replaced if there is an upload request send with the same log id (and we need to allow it as a part of the transactional log upload). This diff fixes an issue where we wouldn't clear the holders of the replaced logs.

Test Plan: Modify the backup integration test to only send logs with id of `1`. Made sure that at the end there were only two blobs with holders -> one for the compaction and one for the last log sent.

Reviewers: kamil, marcin

Reviewed By: kamil

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D11355