[lib] Introduce a handler that processes the ops
Summary:
When an action that doesn't generate ops is dispatched, we can immediately delete it from the queue. Later, in https://linear.app/comm/issue/ENG-7108/resolve-the-promise-after-ops-processing, we're going to listen for an action disappearing from the list.

https://linear.app/comm/issue/ENG-7098/create-a-handler-that-processes-the-ops

Depends on D11284

Test Plan:
Modified update window active action to contain an ID and checked if the handler dispatches actions removing the action from the queue.

Reviewers: kamil, inka, will