[lib] Introduce a handler that processes the ops
Summary:
When an action that doesn't generate ops is dispatched, we can immediately delete it from the queue. In the next diff, we're handling actions that generated some ops.

https://linear.app/comm/issue/ENG-7098/create-a-handler-that-processes-the-ops

Depends on D11284

Test Plan: Modified update window active action to contain an ID and checked if the handler dispatches actions removing the action from the queue.

Reviewers: kamil, inka, will

Reviewed By: kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D11366