[native] Redux migration to patch in `specialRole` field
Summary:
NOTE: Marked as DRAFT to indicate that this diff needs to be tested more thoroughly before landing and is "subject to change." This diff implicitly serves as the Test Plan for D11095 and D11097 which is another reason I'm putting it up in DRAFT state.

---

Depends on D11097

Test Plan:
Will set breakpoints and observe that the shape of data is as expected at each step of the migration.

Will also inspect SQLite DB directly before/after to make sure that things are encoded as expected.

Reviewers: ginsu, tomek, kamil, michal

Reviewed By: tomek, kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D11099