[lib] Handle processing of the ops
Summary:
We're taking the ops from the front of the queue, process it, and dispatch an action that removes it after we're done.

https://linear.app/comm/issue/ENG-7102/reduce-the-action-this-should-filter-the-ops-out-from-the-array

Depends on D11366

Test Plan: Added a sleep to the handler after calling `processDBStoreOperations` and observed in Redux debugger that the ops were processed one-by-one and that the queue length was decreasing.

Reviewers: kamil, inka, will

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D11367