[keyserver][lib] Introduce relationshipActions.FARCASTER_MUTUAL
Summary:
This new relationship action will set `KNOW_OF` for the viewer and the provided users, and create a personal thread. In a later diff, it will also add a robotext message about the Farcaster relationship.

In the future, relationship actions will get moved to the peer-to-peer Olm session, as will the personal thread creation.

Test Plan: Tested this in a later commit by hooking it into `FarcasterDataHandler` (D11387). Can't put up that code quite yet, as it depends on some components from other team members

Reviewers: atul, tomek

Reviewed By: tomek

Differential Revision: https://phab.comm.dev/D11390