[native_rust_library] get_farcaster_users client method
Summary:
native rust client method to call the GetFarcasterUsers RPC. this will be called to get the Comm user info for a user's Farcaster follows

Depends on D11402

Test Plan: later in the stack I add a method to the native IdentityServiceContextProvider which calls this method. This method successfully returns a list of FarcasterUsers.

Reviewers: bartek

Reviewed By: bartek

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D11403