[lib][native] FIDProvider
Summary:
During the Hackathon, this commit enabled us to temporarily store the FID in a context while @will's work on the synced metadata store was in progress.

Since this context is temporary, we considered not diffing it up. Decided against that for two reasons:

1. Putting this diff up allows others to put up diffs while the synced metadata stack (ending in D11424) is pending review.
2. Putting this diff up allows the team to put up existing commits as diffs with minimal modifications.

@will said he'd create a task to cover replacing all `FIDProvider` usages with the synced metadata store. Would be good to mention reverting this diff there as well.

Test Plan: Tested in combination with the rest of the `farcaster` branch

Reviewers: will, atul, ginsu

Reviewed By: will, ginsu

Subscribers: tomek, will

Differential Revision: https://phab.comm.dev/D11425