[identity] check if fid is already taken when registering new user
Summary:
we shouldn't let multiple Comm users associate the same FID with their accounts. So, we need to check on registration if the FID is already taken.

Depends on D11411

Test Plan: tried registering a user with an FID that was already in the users table, it failed as expected. Registration with a new FID (not in DDB) continues to work, though.

Reviewers: bartek

Reviewed By: bartek

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D11412