[identity] on linkFarcasterUser we should return an error if the FID already exists for a different user in the users table
Summary:
LinkFarcasterUser should succeed if the FID is already associated with this user (no-op) but should fail if the FID is already associated with another user. We don't want the same FID associated with multiple Comm users.

Depends on D11412

Test Plan: Tried calling the RPC with an FID already associated with a different user and it failed. Tried calling with an FID associated with this user and it succeeded. Tried calling with an FID associated with no users and it succeeded.

Reviewers: bartek

Reviewed By: bartek

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D11413