[web/native] Add signMessage to olmAPI
Summary:
After rebasing the stack on master now both web and native started using the olm methods to sign messages. Because of that this diff introduces a new method to `olmAPI` -> `signMessage`. On native it's just an alias for `commCoreModule` implementation and on web it proxies to a simple signing function on worker.

Depends on D11373

Test Plan: Called `olmApi.signMessage` on native. Exported the pickled cryptostore from native and imported it on web. Called `olmApi.signMessage` on web. Made sure that the outputs matched (for the same input).

Reviewers: kamil, tomek

Reviewed By: kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D11386