[native] Blocklist for backed up tables
Summary:
[ENG-6918 : Improve backup handling of new tables](https://linear.app/comm/issue/ENG-6918)

Currently tracking if a table is contained in backup happens in two places. Compactions use a SQL query that iterates over tables and runs `DELETE FROM` on backup file. Logs use a vector of table list to watch for. This diff unifies them and has a single blocklist, which informs both compaction and log creation.

Test Plan:
- Created a draft, created a compaction, created another draft (this one is only captured in logs)
- Re-logged, restored backup, made sure both drafts are there
- Added `"drafts"` table to `backedUpTableBlocklist` and repeated the test. This time neither of the logs existed after restore

Reviewers: kamil, marcin

Subscribers: