[lib][web][native] Handle idnentity login in reducers
Summary:
issues: [[https://linear.app/comm/issue/ENG-6598/handle-login-action-call-resetuserspecificstate | ENG-6598]] and [[https://linear.app/comm/issue/ENG-6639/update-specific-reducers-to-handle-login | ENG-6639]]

We want to wipe the state on identity login. More context in [[https://linear.app/comm/issue/ENG-6424/delete-state-on-identity-login-logout-delete-account | ENG-6424]]
We also need to save the CSAT (some further discussion in [[https://linear.app/comm/issue/ENG-7428/discuss-how-to-set-csat-on-native#comment-248c1ba3 | ENG-7428]])
Same as in D10860 we want to set the calendar query in the reducer to use the current dates

Test Plan:
Tested that the state is wiped on normal login. Tested that the state is not wiped on recovery login

Reviewers:
michal, kamil, ginsu
Subscribers: