[native] add the connect farcaster step into the registration flow
Summary:
This diff adds the `ConnectFarcaster` screen into the registration flow. We want to show the connect farcaster screen after the connect etherum screen.

In this diff we also need to add an optional `fid` field to `Params` of all the screens after connect farcaster + `RegistrationServerCallInput` in order to potentially pass an `fid` through the registration flow

Depends on D11442

Test Plan:
Please see the demo video below + tested the entire sign in flow during the hackathon

{F1440207}

Reviewers: atul

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D11461