[keyserver] Don't block innerPerformAuth return on sleep
Summary:
In an earlier version of this stack, in `innerPerformRecovery` I would fall back on legacy login, and without this change the fallback would have to wait on this `sleep`.

This didn't used to matter before because nothing actually waits for `innerPerformAuth` to complete. And since I eventually came to the conclusion that the fallback was a bad idea (we'd be sending user passwords to keyservers), it actually still doesn't matter.

I decided to put this diff up anyways because I think it's probably better for us to avoid blocking the return of `innerPerformAuth` on this `sleep`.

Depends on D11500

Test Plan: Confirm in the earlier version of the stack that when falling back on legacy login, there was no delay due to this `sleep`

Reviewers: tomek, inka

Reviewed By: tomek

Differential Revision: https://phab.comm.dev/D11501