[native] Make thumbnail failed when main upload fails
Summary:
We currently track a video upload and thumbnail upload together, and call `handleUploadFailure` once when either of them fail. We should mark them both as failed, so the thumbnail doesn't appear to be "in progress" from the perspective of `InputState.uploadInProgress`.

This fixes report upload on video message retry, which was previously broken because the thumbnail would appear as in progress to `InputState.uploadInProgress`, which is checked in `canSendReports`.

I considered teasing apart the thumbnail upload so it could tracked on its own, and wouldn't need to be retried if the main upload fails. However this looked to be a more significant investment than I have the time to focus on right now.

Linear task: [ENG-7600](https://linear.app/comm/issue/ENG-7600/media-report-is-not-being-uploaded-when-media-upload-fails)

Depends on D11476

Test Plan: Confirm that when I retry a failed video upload, I see a `SEND_REPORT_SUCCESS` action

Reviewers: atul, bartek

Subscribers: