[sqlite/lib] Split keyserver ops for backed and non-backed up data
Summary:
[ENG-6928 : Return the default data after restoration](https://linear.app/comm/issue/ENG-6928)
Modify the keyserver ops to split the keyserver data into two stringied jsons -> one of them will be backed up and the other won't. When fetching the keyserver data the two tables are joined and the ops merge the data into one. If the non-synced (non-backed up) data isn't there (e.g. after a backup restore), the data will be filled in with defaults.

Depends on D11473

Test Plan:
Run the `keyservers-queries.test.js`. Make sure that keyserver tables are filled in with data correctly. Check that if the non-synced data doesn't have any data, the ops return default values.

Reviewers: kamil, marcin

Subscribers: