[web/native] Migrate keyserverStore to new tables
Summary:
Migrates the current data in one keyserver table to two. It works a bit unusually but the general gist is:
- Copy the data from `keyservers` to `keyserver_synced_data`
- Fetch the data in the migrations (db ops will merge the data from `keyserver` and `keyserver_synced_data` and so return the correct `KeyserverInfo`)
- Save the data in the migration (db ops will now split the data correctly into two tables)

Also adds the non-synced keyservers table to backup blocklist.

Depends on https://phab.comm.dev/D11481

Test Plan:
Make sure the migration runs correctly on both web and native. Compare the results from before the migration and after -> the keyserver infos should be the same (but `urlPrefix` is in a different table).

Reviewers: kamil, marcin

Subscribers: