[lib][web][landing] Don't include injectedWallet in mobile experience
Summary:
The login flow for mobile uses `landing`, while the on for web is contained within `web`.

In earlier versions of Rainbowkit/Wagmi, we could include `injectedWallet` and it would be hidden when viewed from a mobile device.

It's now being included, which isn't desirable for us since our login flow for mobile is rendered instead of a `WebView` on `native`.

This diff lets us use a different set of wallets for `web` and `landing`, with `injectedWallet` being excluded by `landing`.

Depends on D11489

Test Plan: Made sure that "desktop wallet" appears in the desktop SIWE experience and not the mobile SIWE experience

Reviewers: atul, ginsu

Reviewed By: ginsu

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D11490