[lib] Use userStore ops in reducer
Summary: On web we don't persist userStore in redux persist so we can't have a "migration period" where we test if the data in redux persist and sqlite matches and because of that we need to remove the asserts. But this code has been on native for some time now, and seems to work fine so we should be able to remove the safeguards and use data from sqlite directly without any problems.

Test Plan: Play around with the app on web and native, check if there are any errors.

Reviewers: kamil, marcin

Reviewed By: kamil

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D11535