[web] Update type field in DB Media struct
Summary:
This is just to make flow happy in the later diffs. This is safe because we know that we will be the only ones to store `Media` types in the db so we can be sure that the data that we get back has more specific type.

Depends on D11535

Test Plan: flow, tests

Reviewers: kamil, marcin

Reviewed By: kamil

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D11536