[native_rust_library] implement delete_password_user method and move delete methods to a separate file
Summary:
add client methods on native for the new delete_password_user RPCs. this is very similar to identity login, except we don't handle any keys and we don't return anything to the caller.

Depends on D11587

Test Plan: in the next diff, i expose this client method to JS and call this method as a test

Reviewers: bartek!

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D11588