[native] Clear auth metadata on logout
Summary: When a user logs out, we need to clear the auth metadata. This requires adding a function that is responsible for clearing the metadata and emitting new values.

We're setting empty strings as new metadata - which isn't ideal. We're going to improve it as a part of https://linear.app/comm/issue/ENG-7657/improve-auth-metadata-handling.

https://linear.app/comm/issue/ENG-6730/call-clearcommservicesaccesstoken-on-native-logout-and-account

Depends on D11534

Test Plan: Log out and check if an event with empty token was emitted

Reviewers: kamil, inka, varun