[native] migrate threadActivityStore to SQLite
Summary: This migrates threadActivityStore to SQLite

Test Plan:
Test plan is to populate redux persist with threadActivity entries on a lower migration version and then upgrade the version.

This migration will currently fail as is as there is a possible bug with redux persist thread activity not storing `lastPruned`. This is tracked by:
https://linear.app/comm/issue/ENG-7663/missing-lastpruned-field-in-current-thread-activity-redux-persist

By providing a default `lastPruned` value in the reducer, this migration works, but will need to fix ENG-7663 before landing.

Will not land until test plan updated and issue resolved

Reviewers: atul, kamil

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D11616