[lib] Rename `RelationshipRequest` to `TraditionalRelationshipRequest`
Summary:
This is just a straightforward rename. Will want a separate name for the type of RelationshipRequest that includes `userIDs` vs `userIDsToFID`. Initially went with `RelationshipRequestSansFarcaster` or something like that, but thought `TraditionalRelationshipRequest` might be better?

This diff is just rename, the type will be updated so `action` is typed as RelationshipActionSansFarcaster

Test Plan: NA, flow, just a rename

Reviewers: varun, ginsu, tomek, will

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D11602