[lib] Update `FarcasterDataHandler` to construct `FarcasterRelationshipRequest` with `userIDsToFID`
Summary:
We're updating `FarcasterDataHandler` to construct `userIDsToFID` map which will be passed to the `keyserver`. We include `userIDs` and `fids` for both the "target users" as well as current user.

NOTE: The request constructed on the client will currently fail. We need to update both validators and responder on `keyserver` to support this new format.

---

Depends on D11605

Test Plan:
Add log statements to `userIDsToFID`, `userIDsToFIDIncludingCurrentUser`, and `createThreadsAndRobotextForFarcasterMutuals` and ensure that data is structured as expected.

(made some tweaks to the code here, so will go through again before landing)

Reviewers: ginsu, varun, will, tomek

Reviewed By: ginsu, tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D11606