[lib/native] introduce connect farcaster alert handler
Summary:
This diff introduces `ConnectFarcasterAlertHandler` component which handles the mechanism to prompt the user to connect to farcaster.

I needed to introduce a new handler because `FarcasterDataHandler` lives in `lib` and I needed to be able to use the `useNavigation` hook.

Linear task: https://linear.app/comm/issue/ENG-7525/extend-farcaster-data-handler-to-handle-navigating-to-the-connect + https://linear.app/comm/issue/ENG-7623/introduce-shouldskipconnectfarcasteralert

Depends on D11601

Test Plan:
Please see the demo video below

{F1507626}

Reviewers: atul, inka

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D11613