[lib] Stop attempting recovery when client version unsupported
Summary:
issue: [[https://linear.app/comm/issue/ENG-6125/if-client-version-unsupported-skip-fetchnewcookiefromnativecredentials | ENG-6125]]
If the client got `client_version_unsupported`, a recovery will not help them, so we should skip it. We should just set the session change sent by the server, and a `connectionIssue` so that other places in the code know what the problem was.

Test Plan:
Tested that if the auth keyserver sends `client_version_unsupported`, then the user is logged out. Checked that a `connectionIssue` was set, and that triggered the logout.
I tested in the multikeyserver setup, that if the second keyserver sends `client_version_unsupported`, then the client is not logged out, and a popup shows up. For now, this results in the screen being spammed with popups, because [[https://linear.app/comm/issue/ENG-5913/check-all-places-that-select-all-keyservers-from-keyserverstore#comment-3cbd4a93 | ENG-5913]] was not implemented yet (only in case of the secondary keyserver).

Reviewers:
tomek, ginsu
Subscribers: